+++ src/config/prelude.rs
@@ -11,7 +11,7 @@ use himalaya_lib::ImapConfig;
 #[cfg(feature = "notmuch-backend")]
 use himalaya_lib::NotmuchConfig;
 
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(remote = "SmtpConfig")]
 struct SmtpConfigDef {
     #[serde(rename = "smtp-host")]
@@ -31,7 +31,7 @@ struct SmtpConfigDef {
 }
 
 #[cfg(feature = "imap-backend")]
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(remote = "ImapConfig")]
 pub struct ImapConfigDef {
     #[serde(rename = "imap-host")]
@@ -56,41 +56,39 @@ pub struct ImapConfigDef {
     pub watch_cmds: Option<Vec<String>>,
 }
 
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "MaildirConfig")]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(remote = "MaildirConfig", rename_all = "kebab-case")]
 pub struct MaildirConfigDef {
     #[serde(rename = "maildir-root-dir")]
     pub root_dir: PathBuf,
 }
 
 #[cfg(feature = "notmuch-backend")]
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "NotmuchConfig")]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(remote = "NotmuchConfig", rename_all = "kebab-case")]
 pub struct NotmuchConfigDef {
     #[serde(rename = "notmuch-db-path")]
     pub db_path: PathBuf,
 }
 
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "Option<EmailTextPlainFormat>")]
-pub enum EmailTextPlainFormatOptionDef {
-    #[serde(with = "EmailTextPlainFormatDef")]
-    Some(EmailTextPlainFormat),
-    #[default]
-    None,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "EmailTextPlainFormat", rename_all = "snake_case")]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(
+    remote = "EmailTextPlainFormat",
+    tag = "type",
+    content = "width",
+    rename_all = "kebab-case"
+)]
 pub enum EmailTextPlainFormatDef {
+    #[default]
     Auto,
     Flowed,
     Fixed(usize),
 }
 
-#[derive(Debug, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "EmailSender", tag = "sender", rename_all = "snake_case")]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(remote = "EmailSender", tag = "sender", rename_all = "kebab-case")]
 pub enum EmailSenderDef {
+    #[default]
     None,
     #[serde(with = "SmtpConfigDef")]
     Smtp(SmtpConfig),
@@ -98,26 +96,17 @@ pub enum EmailSenderDef {
     Sendmail(SendmailConfig),
 }
 
-#[derive(Debug, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "SendmailConfig")]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(remote = "SendmailConfig", rename_all = "kebab-case")]
 pub struct SendmailConfigDef {
     #[serde(rename = "sendmail-cmd")]
     cmd: String,
 }
 
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "Option<EmailHooks>")]
-pub enum EmailHooksOptionDef {
-    #[serde(with = "EmailHooksDef")]
-    Some(EmailHooks),
-    #[default]
-    None,
-}
-
 /// Represents the email hooks. Useful for doing extra email
 /// processing before or after sending it.
-#[derive(Debug, Default, Clone, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(remote = "EmailHooks")]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
+#[serde(remote = "EmailHooks", rename_all = "kebab-case")]
 pub struct EmailHooksDef {
     /// Represents the hook called just before sending an email.
     pub pre_send: Option<String>,