+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed config deserialization issue with `email-hooks` and
+  `email-reading-format`.
+
 ## [0.7.1] - 2023-02-14
 
 ### Added